# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Modern Men Hair Salon API
  description: |
    Complete API specification for the Modern Men Hair Salon management system.

    ## Features
    - Customer management with loyalty programs
    - Appointment scheduling and management
    - Service catalog with pricing
    - Staff management and scheduling
    - Real-time notifications
    - Analytics and reporting

    ## Authentication
    All API endpoints require authentication via JWT tokens or session cookies.
    Include the Authorization header with your Bearer token for authenticated requests.
  version: 1.0.0
  contact:
    name: Modern Men Support
    email: support@modernmen.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.modernmen.com/v1
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /services:
    get:
      summary: Get services with filtering
      description: Retrieve services with advanced filtering and search capabilities
      tags:
        - Services
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: category
          in: query
          description: Filter by service category
          schema:
            type: string
        - name: isActive
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - name: minPrice
          in: query
          description: Minimum price filter
          schema:
            type: number
            minimum: 0
        - name: maxPrice
          in: query
          description: Maximum price filter
          schema:
            type: number
            minimum: 0
        - name: search
          in: query
          description: Search in name, description, or category
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field and direction
          schema:
            type: string
            enum: [name, -name, price, -price, createdAt, -createdAt]
        - name: limit
          in: query
          description: Number of services per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        '200':
          description: Services retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  docs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
                  totalDocs:
                    type: integer
                  limit:
                    type: integer
                  totalPages:
                    type: integer
                  page:
                    type: integer
                  meta:
                    type: object
                    properties:
                      filters:
                        type: object
                      availableCategories:
                        type: array
                        items:
                          type: string

    post:
      summary: Create a new service
      description: Create a new service with duplicate prevention and validation
      tags:
        - Services
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
                - duration
                - category
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  maxLength: 1000
                price:
                  type: number
                  minimum: 0
                  maximum: 10000
                duration:
                  type: integer
                  minimum: 15
                  maximum: 480
                category:
                  type: string
                  minLength: 1
                  maxLength: 50
                isActive:
                  type: boolean
                  default: true
                image:
                  type: string
                  format: uri
                features:
                  type: array
                  items:
                    type: string
                tenant:
                  type: string
      responses:
        '201':
          description: Service created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Service'
                  message:
                    type: string
                    example: "Service created successfully"
...
components:
  schemas:
...
    Service:
      type: object
      properties:
        id:
          type: string
          description: Unique service identifier
        name:
          type: string
        description:
          type: string
        price:
          type: number
        duration:
          type: integer
          description: Duration in minutes
        category:
          type: string
        isActive:
          type: boolean
        image:
          type: string
          format: uri
        features:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie for authenticated users